Uploaded Search Dictionary Files
Watched a 30min tutorial about creating a search dictionary for Scrabble, but the methods seemed very basic for me.  Therefore, I removed a lot of the repetitive code from the tutorial and practiced using a header file.  I didn't need to create it, but did it for practice.  I'll definitely edit it down the line. 
This is my first post of many to follow.  Feel free to rip at the bad code as I am aiming to only go up from here with my coding. 

Thx in advanced.